DAP - create authentication separate folder
